+++ lab1/CMakeLists.txt
@@ -3,4 +3,5 @@ project(lab1 C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(lab1 main.c)
+add_executable(lab1 main.c
+        find_max_incresing_subsequence.h)
