+++ lab1/cmake-build-debug/.ninja_log
@@ -1,7 +1,14 @@
 # ninja log v5
-0	33	1710615252058251863	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
-34	75	1710615252099118951	lab1	2327f7dc1dd32b0d
-0	40	1710617168995363512	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
-40	83	1710617169038708625	lab1	2327f7dc1dd32b0d
-0	45	1710617964599635609	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
-45	85	1710617964639503163	lab1	2327f7dc1dd32b0d
+32	72	1710620049345432829	lab1	2327f7dc1dd32b0d
+0	32	1710620049304668995	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+1	41	1710621203269597471	build.ninja	36760719a1742074
+1	37	1710622424109940460	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+37	79	1710622424152394687	lab1	2327f7dc1dd32b0d
+0	32	1710622543867781717	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+32	69	1710622543907014808	lab1	2327f7dc1dd32b0d
+0	41	1710622601155905676	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+41	82	1710622601195842391	lab1	2327f7dc1dd32b0d
+1	41	1710622623928850472	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+42	83	1710622623969925798	lab1	2327f7dc1dd32b0d
+0	36	1710622652939508340	CMakeFiles/lab1.dir/main.c.o	11846846e0c5a353
+36	74	1710622652977055405	lab1	2327f7dc1dd32b0d
