+++ lab1/cmake-build-debug/CMakeFiles/CMakeConfigureLog.yaml
@@ -192,3 +192,39 @@ events:
       
       
 ...
+
+---
+events:
+  -
+    kind: "try_compile-v1"
+    backtrace:
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/Internal/CheckSourceCompiles.cmake:101 (try_compile)"
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/CheckCSourceCompiles.cmake:52 (cmake_check_source_compiles)"
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/FindThreads.cmake:97 (CHECK_C_SOURCE_COMPILES)"
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/FindThreads.cmake:163 (_threads_check_libc)"
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/CMakeFindDependencyMacro.cmake:76 (find_package)"
+      - "/opt/homebrew/lib/cmake/GTest/GTestConfig.cmake:29 (find_dependency)"
+      - "/Applications/CLion.app/Contents/bin/cmake/mac/aarch64/share/cmake-3.27/Modules/FindGTest.cmake:194 (find_package)"
+      - "CMakeLists.txt:10 (find_package)"
+    directories:
+      source: "/Users/askoritan/programming/algds2024/lab1/cmake-build-debug/CMakeFiles/CMakeScratch/TryCompile-1yBhFH"
+      binary: "/Users/askoritan/programming/algds2024/lab1/cmake-build-debug/CMakeFiles/CMakeScratch/TryCompile-1yBhFH"
+    cmakeVariables:
+      CMAKE_C_FLAGS: ""
+      CMAKE_C_FLAGS_DEBUG: "-g"
+      CMAKE_EXE_LINKER_FLAGS: ""
+      CMAKE_OSX_ARCHITECTURES: ""
+      CMAKE_OSX_DEPLOYMENT_TARGET: ""
+      CMAKE_OSX_SYSROOT: "/Library/Developer/CommandLineTools/SDKs/MacOSX14.2.sdk"
+    buildResult:
+      variable: "CMAKE_HAVE_LIBC_PTHREAD"
+      cached: true
+      stdout: |
+        Change Dir: '/Users/askoritan/programming/algds2024/lab1/cmake-build-debug/CMakeFiles/CMakeScratch/TryCompile-1yBhFH'
+        
+        Run Build Command(s): /Applications/CLion.app/Contents/bin/ninja/mac/aarch64/ninja -v cmTC_af8b1
+        [1/2] /Library/Developer/CommandLineTools/usr/bin/cc -DCMAKE_HAVE_LIBC_PTHREAD  -std=gnu11 -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX14.2.sdk -fcolor-diagnostics -MD -MT CMakeFiles/cmTC_af8b1.dir/src.c.o -MF CMakeFiles/cmTC_af8b1.dir/src.c.o.d -o CMakeFiles/cmTC_af8b1.dir/src.c.o -c /Users/askoritan/programming/algds2024/lab1/cmake-build-debug/CMakeFiles/CMakeScratch/TryCompile-1yBhFH/src.c
+        [2/2] : && /Library/Developer/CommandLineTools/usr/bin/cc -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX14.2.sdk -Wl,-search_paths_first -Wl,-headerpad_max_install_names  CMakeFiles/cmTC_af8b1.dir/src.c.o -o cmTC_af8b1   && :
+        
+      exitCode: 0
+...
