+++ lab1/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Mar 16 22:46 MSK
+Start testing: Mar 16 23:57 MSK
 ----------------------------------------------------------
-End testing: Mar 16 22:46 MSK
+End testing: Mar 16 23:57 MSK
